+++ code/game/objects/structures/lattice.dm
@@ -1,52 +1,56 @@
 /obj/structure/lattice
 	name = "lattice"
 	desc = "A lightweight support lattice."
-	icon = 'icons/obj/structures.dmi'
-	icon_state = "latticefull"
-	density = 0
-	anchored = 1.0
+	icon = 'icons/obj/smoothlattice.dmi'
+	icon_state = "lattice0"
+	density = FALSE
+	anchored = TRUE
 	w_class = ITEM_SIZE_NORMAL
-	plane = -18
+	plane = ABOVE_PLATING_PLANE
 	layer = LATTICE_LAYER
+	color = COLOR_STEEL
+	var/init_material = MATERIAL_STEEL
 	//	obj_flags = OBJ_FLAG_CONDUCTIBLE
 	mass = 5
 	max_health = 40
 
-/obj/structure/lattice/Initialize()
+/obj/structure/lattice/get_material()
+	return material
+
+/obj/structure/lattice/Initialize(mapload, var/new_material)
 	. = ..()
+	if(!(istype(src.loc, /turf/space) || istype(src.loc, /turf/simulated/open)))
+		return INITIALIZE_HINT_QDEL
+	if(!new_material)
+		new_material = init_material
+	material = SSmaterials.get_material_by_name(new_material)
+	if(!istype(material))
+		return INITIALIZE_HINT_QDEL
+
+	SetName("[material.display_name] lattice")
+	desc = "A lightweight support [material.display_name] lattice."
+	color =  material.icon_colour
+
 	for(var/obj/structure/lattice/LAT in loc)
 		if(LAT != src)
 			log_debug("Found multiple lattices at '[log_info_line(loc)]'. Deleting extra lattices!")
 			qdel(LAT)
-	icon = 'icons/obj/smoothlattice.dmi'
-	icon_state = "latticeblank"
-	updateOverlays()
-	for (var/dir in GLOB.cardinal)
-		var/obj/structure/lattice/L
-		if(locate(/obj/structure/lattice, get_step(src, dir)))
-			L = locate(/obj/structure/lattice, get_step(src, dir))
-			L.updateOverlays()
+	update_icon()
+	if(!mapload)
+		update_neighbors()
 
 /obj/structure/lattice/Destroy()
-	for (var/dir in GLOB.cardinal)
-		var/obj/structure/lattice/L
-		if(locate(/obj/structure/lattice, get_step(src, dir)))
-			L = locate(/obj/structure/lattice, get_step(src, dir))
-			L.updateOverlays(src.loc)
+	var/turf/old_loc = get_turf(src)
 	. = ..()
+	if(old_loc)
+		update_neighbors(old_loc)
 
-/obj/structure/lattice/ex_act(severity)
-	switch(severity)
-		if(1.0)
-			qdel(src)
-			return
-		if(2.0)
-			qdel(src)
-			return
-		if(3.0)
-			return
-		else
-	return
+/obj/structure/lattice/proc/update_neighbors(var/location = loc)
+	for (var/dir in GLOB.cardinal)
+		var/obj/structure/lattice/L = locate(/obj/structure/lattice, get_step(location, dir))
+		if(L)
+			L.update_icon()
+
 
 /obj/structure/lattice/attackby(obj/item/C as obj, mob/user as mob)
 
@@ -58,11 +62,10 @@
 		var/obj/item/weapon/tool/weldingtool/WT = C
 		if(WT.remove_fuel(0, user))
 			to_chat(user, "<span class='notice'>Slicing lattice joints ...</span>")
-		new /obj/item/stack/rods(loc)
+		new /obj/item/stack/material/rods(loc, 1, material.name)
 		qdel(src)
-
-/**	if (istype(C, /obj/item/stack/rods))
-		var/obj/item/stack/rods/R = C
+	if (istype(C, /obj/item/stack/material/rods))
+		var/obj/item/stack/material/rods/R = C
 		if(R.use(2))
 			src.alpha = 0
 			playsound(src, 'sound/weapons/Genhit.ogg', 50, 1)
@@ -71,26 +74,15 @@
 			return
 		else
 			to_chat(user, "<span class='notice'>You require at least two rods to complete the catwalk.</span>")
-			return
-**/
-	return
 
-/obj/structure/lattice/proc/updateOverlays()
-	//if(!(istype(src.loc, /turf/space)))
-	//	qdel(src)
-	spawn(1)
-		overlays = list()
-
-		var/dir_sum = 0
-
-		var/turf/T
-		for (var/direction in GLOB.cardinal)
-			T = get_step(src, direction)
-			if(locate(/obj/structure/lattice, T) || locate(/obj/structure/catwalk, T))
+/obj/structure/lattice/on_update_icon()
+	var/dir_sum = 0
+	for (var/direction in GLOB.cardinal)
+		var/turf/T = get_step(src, direction)
+		if(locate(/obj/structure/lattice, T) || locate(/obj/structure/catwalk, T))
+			dir_sum += direction
+		else
+			if(!(istype(get_step(src, direction), /turf/space)) && !(istype(get_step(src, direction), /turf/simulated/open)))
 				dir_sum += direction
-			else
-				if(!(istype(get_step(src, direction), /turf/space)) && !(istype(get_step(src, direction), /turf/simulated/open)))
-					dir_sum += direction
 
-		icon_state = "lattice[dir_sum]"
-		return
+	icon_state = "lattice[dir_sum]"
